review: address PR #177 feedback
- cwd now REQUIRED in both ExecuteTestGenerationInputSchema and
  ExecuteReplayInputSchema. Drops the `?? process.cwd()` default in
  tool-registry.ts — ambiguity between an explicit cwd and the MCP server's
  startup cwd could cause the lock to mis-key and split one logical change
  into two competing holders. Skill always passes cwd explicitly now.
- Move lock constants out to local-execution-lock-constants.ts (logic file
  stays logic-only).
- Move lock types out to local-execution-lock-types.ts (ILockState,
  ILocalExecutionLockHandle, ITryAcquireOutcome).
- Move worktree-creation details out of SKILL.md to
  _shared/pr-branch-worktree.md; SKILL.md just references.
